// apps/buyer-platform/src/common/enum/document.enum.ts
export enum DocumentType {
  COPY_OF_ID_OR_PASSPORT_AUTHORIZED_PERSON = 'COPY_OF_ID_OR_PASSPORT_AUTHORIZED_PERSON',
  COPY_OF_HOUSE_REGISTRATION_AUTHORIZED_PERSON = 'COPY_OF_HOUSE_REGISTRATION_AUTHORIZED_PERSON',
  ID_CARD_WITH_PERSON = 'ID_CARD_WITH_PERSON',
  POWER_OF_ATTORNEY = 'POWER_OF_ATTORNEY',
  COPY_OF_ID_OR_PASSPORT_ATTORNEY_GIVER = 'COPY_OF_ID_OR_PASSPORT_ATTORNEY_GIVER',

  COPY_OF_COMPANY_REGISTRATION = 'COPY_OF_COMPANY_REGISTRATION',
  COPY_OF_VAT_REGISTRATION = 'COPY_OF_VAT_REGISTRATION',
  TRADEMARK = 'TRADEMARK',
  PHOTO_OF_COMPANY_OR_PROJECT = 'PHOTO_OF_COMPANY_OR_PROJECT',
  COPY_OF_CERTIFIED_COMMERCIAL_REGISTRATION_WITH_SEAL = 'COPY_OF_CERTIFIED_COMMERCIAL_REGISTRATION_WITH_SEAL', // เฉพาะบุคคลจดทะเบียนพาณิชย์

  COPY_OF_BANK_ACCOUNT_RECEIVING_PAYMENTS = 'COPY_OF_BANK_ACCOUNT_RECEIVING_PAYMENTS',

  OTHERS = 'OTHERS',
}